refactor: move order placement to order route
