+++ api/admin/adminMiddleware.js
@@ -1,5 +1,4 @@
 export default function (req, res, next) {
-  console.log(req.user.isAdmin);
   if (!req.user.isAdmin) return res.status(403).send('Access denied');
   next();
 }
