+++ api/order/orderRoute.js
@@ -1,9 +1,15 @@
 import express from 'express';
+import { check, validationResult } from 'express-validator';
 import Order from './Order.js';
 import User from '../user/User.js';
 import Ticket from '../ticket/Ticket.js';
 import authMiddleware from '../authentication/authMiddleware.js';
 import adminMiddleware from '../admin/adminMiddleware.js';
+import orderMiddleware from './orderMiddleware.js';
+import transporter from '../../mail/transporter.js';
+import getMailOptions from '../../mail/mailOptions.js';
+import Seat from '../seat/Seat.js';
+import Screening from '../screening/Screening.js';
 
 const router = express.Router();
 
@@ -11,8 +17,8 @@ router
   .route('/')
   // @route GET api/orders
   // @description Get all orders
-  // @access public
-  .get(async (req, res) => {
+  // @access admin
+  .get(authMiddleware, adminMiddleware, async (req, res) => {
     try {
       const orders = await Order.find({});
       res.status(200).json(orders);
@@ -21,49 +27,162 @@ router
     }
   })
   // @route POST api/orders
-  // @description Create an orders
-  // @access admin
-  .post(authMiddleware, adminMiddleware, async (req, res) => {
-    // eslint-disable-next-line object-curly-newline
-    const { userId, email, status, tickets } = req.body;
-    try {
-      const user = await User.findById(userId);
-      if (user === undefined) {
-        res.status(404).json({
-          error: `Cannot find user with id: ${req.params.id}'`,
-        });
-        return;
+  // @description Create an order
+  // @access public
+  // @example req.body:
+  // @{
+  // @  "status": "pending",
+  // @  "tickets": [[0,0],[1,0]],
+  // @  "screening": "604cd23fc5d64540d07eaece"
+  // @}
+  .post(
+    orderMiddleware,
+    check('status', 'Status is required')
+      .trim()
+      .notEmpty()
+      .isString(),
+    check('screening', 'Screening is required')
+      .trim()
+      .notEmpty()
+      .isString(),
+    check('email', 'Please include a valid email')
+      .isEmail()
+      .trim()
+      .isLength({ max: 255 })
+      .normalizeEmail(),
+    check('tickets', 'Ticket is required').notEmpty().isArray(),
+    async (req, res) => {
+      const errors = validationResult(req);
+      if (!errors.isEmpty()) {
+        return res.status(400).json({ errors: errors.array() });
       }
-      const order = new Order({
-        user: userId,
-        email,
-        status,
-      });
-      await order.save();
-      tickets.forEach(async (ticket) => {
-        const newTicket = new Ticket({
-          screening: ticket.screeningId,
-          row: ticket.row,
-          column: ticket.column,
-          order: order.id,
+
+      try {
+        const { email, status, tickets, screening } = req.body;
+        const { isAuthenticated, user } = req;
+        const userPlacingOrder = isAuthenticated
+          ? await User.findById(user.id).select('email orders')
+          : undefined;
+
+        // Get screening for which the order was placed
+        const screeningChosen = await Screening.findById(screening)
+          .select('cinemaHall')
+          .populate({ path: 'cinemaHall' });
+        // If the screening doesn't exist send 404
+        if (!screeningChosen) {
+          return res.status(404).send('Screening not found');
+        }
+        // Get tickets that exist for this screening
+        const existingTickets = await Ticket.find({
+          screening: screeningChosen.id,
+        }).populate({
+          path: 'seat',
+          model: 'Seat',
         });
-        await newTicket.save();
-        return ticket.id;
-      });
-      await order.save();
-      res.status(200).json({ message: order.id });
-    } catch (e) {
-      res.status(400).send(e);
-    }
-  });
+        // Prepare seats tuples that are occupied
+        const occupiedSeats = existingTickets.map(({ seat }) => [
+          seat.row,
+          seat.column,
+        ]);
+        // Check if any of the seats that the order is placed for is occupied
+        // eslint-disable-next-line arrow-body-style
+        const isSpaceOccupied = ([row, column]) => {
+          if (occupiedSeats.length === 0) return false;
+          return occupiedSeats.some(
+            ([occupiedSeatRow, occupiedSeatColumn]) => {
+              let result = false;
+              if (
+                // eslint-disable-next-line operator-linebreak
+                occupiedSeatRow === row &&
+                occupiedSeatColumn === column
+              ) {
+                result = true;
+              }
+              return result;
+            },
+          );
+        };
+        const areEmpty = tickets.every(
+          (ticket) => !isSpaceOccupied(ticket),
+        );
+        if (!areEmpty) {
+          return res.status(404).send('Seats are not empty');
+        }
+        const seats = await Promise.all(
+          tickets.map(([rowNr, columnNr]) => {
+            const seat = Seat.findOne({
+              hall: screeningChosen.cinemaHall.id,
+              row: rowNr,
+              column: columnNr,
+            });
+            return seat;
+          }),
+        );
+        if (!seats) {
+          return res.status(404).send('Seats not found');
+        }
+
+        // If all data are ok, create order
+        const order = new Order({
+          email,
+          status,
+        });
+        if (isAuthenticated) {
+          order.user = user.id;
+        }
+
+        // Generate all tickets and their ids
+        const ticketsIds = await order.createOrdersDependencies(
+          seats,
+          screening,
+          order,
+          (err) => {
+            if (err) {
+              return res.status(400).json({
+                msg: 'Can not generate tickets for this order',
+              });
+            }
+          },
+        );
+
+        // Add ticket ids as reference to the order
+        ticketsIds.forEach((ticketId) => {
+          order.tickets.push(ticketId);
+        });
+
+        // Save order in mongo db and update users orders if user is authenticated
+        await order.save();
+        if (userPlacingOrder) {
+          userPlacingOrder.orders.push(order.id);
+          await userPlacingOrder.save();
+        }
+
+        // Send email with notifications
+        const emailOptions = getMailOptions(
+          email,
+          'Order Placed',
+          `Your order number: ${order.id} was successfully placed`,
+        );
+        transporter.sendMail(emailOptions, (error, info) => {
+          if (error) {
+            console.log(error);
+          } else {
+            console.log(`Email sent: ${info.response}`);
+          }
+        });
+        res.status(200).json({ msg: 'POSTED', order: order });
+      } catch (e) {
+        res.status(400).send({ msg: 'Error placing order' });
+      }
+    },
+  );
 
 router
   .route('/:id')
   // @route GET api/orders/id
   // @description Get an order
-  // @access user
-  .get(authMiddleware, async (req, res) => {
-    // TODO: for admin and user who purchased only? to change for all not public routes
+  // @access admin
+  .get(authMiddleware, adminMiddleware, async (req, res) => {
     const order = await Order.findById(req.params.id);
     try {
       if (order === undefined) {
@@ -77,58 +196,49 @@ router
       res.status(400).send(e);
     }
   })
-  // @route PUT api/orders/id
-  // @description Update an order
-  // @access admin
-  .put(authMiddleware, adminMiddleware, async (req, res) => {
-    const { userId, email, status } = req.body;
-    const order = await Order.findById(req.params.id);
-    try {
-      if (order === undefined) {
-        res.status(404).json({
-          error: `Cannot find order with id: ${req.params.id}'`,
-        });
-        return;
-      }
-      if (userId !== undefined) {
-        const user = await User.findById(userId);
-        if (user === undefined) {
-          res.status(400).json({
-            error: `Cannot find user with id: ${req.params.id}'`,
-          });
-          return;
-        }
-        order.user = userId;
-      }
-      if (email !== undefined) {
-        order.email = email;
-      }
-      if (status !== undefined) {
-        order.status = status;
-      }
-      await order.save();
-      res
-        .status(200)
-        .json({ message: 'Order updated successfully', order });
-    } catch (e) {
-      res.status(400).send(e);
-    }
-  })
+
   // @route DELETE api/orders/id
   // @description Delete an order
   // @access admin
   .delete(authMiddleware, adminMiddleware, async (req, res) => {
     try {
-      // check if order exists
-      // check order.findById -> order.tickets  -> delete tickets -> delete order
-      const ticket = await Ticket.findByIdAndDelete(req.params.id);
-      if (ticket === undefined) {
-        res.status(404).json({
-          error: `Cannot find ticket with id: ${req.params.id}'`,
+      // Find order and delete
+      // Tickets are removed automatically by orderSchema.post('remove')
+      const order = await Order.findByIdAndDelete(req.params.id);
+      if (!order) {
+        return res.status(404).json({
+          error: `Cannot find order with id: ${req.params.id}'`,
         });
-        return;
       }
-      res.status(204).json(ticket);
+
+      // Get email of the client and send notification
+      const { email } = order;
+      const emailOptions = getMailOptions(
+        email,
+        'Order Deleted',
+        `Your order number: ${order.id} was successfully deleted`,
+      );
+      transporter.sendMail(emailOptions, (error, info) => {
+        if (error) {
+          console.log(error);
+        } else {
+          console.log(`Email sent: ${info.response}`);
+        }
+      });
+
+      const userId = order.user;
+      if (userId) {
+        const user = await User.findByIdAndUpdate(userId, {
+          $pull: {
+            orders: {
+              _id: req.params.id,
+            },
+          },
+        });
+        await user.save();
+      }
+
+      res.status(204).json(order);
     } catch (e) {
       res.status(400).send(e);
     }
